python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "granola-sync"
version = "1.0.0"
description = "Export Granola meetings to Google Drive as .docx files"
readme = {text = "Export Granola meetings to Google Drive as .docx files", content-type = "text/plain"}
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Mahmoud Salim" },
]
dependencies = [
    "python-docx>=1.1.0",
    "requests>=2.31.0",
    "lxml>=5.0.0",
]

[project.scripts]
granola-sync = "granola_sync.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/granola_sync"]

[tool.pytest.ini_options]
testpaths = ["tests"]
